Switch from EventSource to WebSocket
Scaling EventSoure was going to be annoying, since there was no
guarantee that a write request would end up on the same backend server,
meaning that either requests would have to make it to the same backend
or they'd have to be pushed out via SNS/SQS to the other nodes.  With
WebSocket, writes just happen over the same connection. Debugging in
cURL is nice, but that's what wscat is for.

No more connection ids (for now, though they might return if reconnects
become taxing), and delete requests now take a subscriptions array
instead of just an apiKey/topic.

Some errors for write requests currently close the connection, but some
of those might change to be non-fatal.