Include 'delay' in global topic notification
Instead of waiting to send global topic notifications and letting the
client pick a delay, include a randomly chosen delay in the notification
that clients should obey. It's not particularly appropriate to have this
in stream-server itself, but it's much better to be able to adjust the
delay server-side than to leave it up to hard-coded update periods in
the client.